requirements.txt
Faker==14.2.0
psycopg2==2.9.3
python-decouple==3.6
